breaking out lexer crate for maintainability
