+++ crates/runmat-lexer/src/extras.rs
@@ -1,0 +1,5 @@
+#[derive(Default, Clone, Copy)]
+pub struct LexerExtras {
+    pub last_was_value: bool,
+    pub line_start: bool,
+}
